requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-docx
Pillow
